# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Medicinal Herbs API",
    "description": "API for managing medicinal herbs and products, including CRUD operations.",
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": ["username", "email", "password", "role"],
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true,
            "description": "Unique identifier of the user"
          },
          "username": {
            "type": "string",
            "description": "Username of the user",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email of the user",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password of the user",
            "example": "password123"
          },
          "role": {
            "type": "string",
            "description": "Role of the user",
            "example": "researcher"
          },
          "is_admin": {
            "type": "boolean",
            "description": "Whether the user is an admin",
            "example": false
          }
        }
      },
...
      "Product": {
        "type": "object",
        "required": ["name", "price", "seller_id"],
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true,
            "description": "Unique identifier of the product"
          },
          "name": {
            "type": "string",
            "description": "Name of the product",
            "example": "Basil Essential Oil"
          },
          "description": {
            "type": "string",
            "description": "Description of the product",
            "example": "Pure basil essential oil extracted from organic basil leaves."
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Price of the product",
            "example": 19.99
          },
          "stock": {
            "type": "integer",
            "description": "Stock quantity of the product",
            "example": 100
          },
          "seller_id": {
            "type": "integer",
            "description": "ID of the seller (user) who created the product",
            "example": 1
          }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "description": "Register a new user with the provided details.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log in a user",
        "description": "Log in a user and return a JWT token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Invalid email or password"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/auth/products": {
      "post": {
        "tags": ["Auth"],
        "summary": "Add a new product",
        "description": "Add a new product to the marketplace.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product added successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/admin/login": {
      "post": {
        "tags": ["Admin"],
        "summary": "Log in an admin",
        "description": "Log in an admin and return a JWT token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Invalid email or password"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/admin/herbs": {
      "get": {
        "tags": ["Admin"],
        "summary": "View all herbs",
        "description": "View all herbs in the database.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Herbs retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}